Update
1. 拒绝`tun0`侧主动请求的连接.
2. 加入`EXIT`环境变量.

+++ docker-root/usr/local/bin/start-sangfor.sh
@@ -12,6 +12,9 @@ do
 	
 	# 下面这行代码启动 EasyConnect 的前端。
 	/usr/share/sangfor/EasyConnect/EasyConnect --enable-transparent-visuals --disable-gpu
+	
+	# 是否自动重启
+	[ -n "$EXIT" ] && exit
 
 	# 清除的残余进程，它们可能会妨碍下次的启动。
 	killall CSClient svpnservice
